src/auth_service/requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic
python-jose
httpx
dotenv
psycopg2-binary
passlib
bcrypt